requirements.txt
transformers
torch
llama-cpp-python
